Unit tests part 2
