+++ lib/process-services/src/lib/mock/public-api.ts
@@ -26,3 +26,5 @@ export * from './task/task-details.mock';
 export * from './task/task-details.component.mock';
 export * from './task/task-list.mock';
 export * from './task/tasklist-service.mock';
+export * from './process/process-filters.mock';
+export * from './task/task-filters.mock';
